feat: qtt generate/validate execution plans
This patch extends our query validation tests to generate, execute, and
validate execution plans.

TestExecutor/TestExecutorUtil have been extended to support executing from
a plan as an alternative to a statement, if included in the TestCase spec.
In both cases, TestExecutorUtil first builds an Iterable that yields
instances of ConfiguredKsqlPlan. If the plans are included in the TestCase,
the Iterable just iterates over the supplied plans. If not, the Iterable
builds each statement into a plan.

Execution plans are derived from the qtt test cases and checked into the
repository under `ksql-functional-tests/src/test/resources/historical_plans`.
The plans are organized as follows:
`<test-case>`: all the plans for a given test case
`<test-case>/<version>_<timestamp>`: the plan for a given test case at a
                                     point in time
`<test-case>/<version>_<timestamp>`/spec.json - a json containing the config,
                                                plan, and any generated schemas.
`<test-case>/<version>_<timestamp>`/topology - the streams topology description.

A new test loader called PlannedTestLoader iterates over all the plans and
builds instances of TestCase out of them.

To ensure we include new query plans over time, this patch includes a test
called PlannedTestUpToDateTest, that iterates over each qtt test case and
ensures that we have persisted the current plan that the engine builds.

Finally, this patch includes PlannedTestGenerator which you can use to
generate any new plans that need to be generated.