test: enhance test framework to support primitive keys (#4201)
* test: enhance test framework to support primitive keys

Enhanced the test framework to work with test cases with `BIGINT` and `DOUBLE` keys. This was previously failing due to Jackson deserializing the JSON describing the expected values into to `INT` and `DECIMAL`.

To achieve this I've started to tidy up the mess in the test framework. Specifically, we no longer track the serde to use in the `Topic` class, as this is constructed when parsing the test cases. Doing so was problematic as it was hard to know what serde should be used just from the test case text.  The serde to use is now determined by looking at the format of sources in the metastore once the SQL statements have been executed.

Also includes some additional test cases for pull queries working with primitive keys, and pulling out the code that determines the format / serde to use for a topic into a `TopicInfoCache`, as this is not shared by `QTT` and `RQTT` tests.