chore: fix TestDataProvider's to have correct key type
A couple of the `TestDataProvider` implementations have a `WITH(KEY)` field that has a non-`STRING` SQL type, but the SQL type of `ROWKEY` was still `STRING`.
This is not incorrect: the with-key field must have the same SQL type as the actual key, i.e. `ROWKEY`.

This change fixes the `TestDataProvider`s and any associated tests.  It also refactors the providers and the code tests use to produce and consume messages to clean this up and remove duplicate functionality.