feat: improve the semantics around decimal/floating point literals
BREAKING CHANGE: standalone literals that used to be doubles may now be
interpreted as BigDecimal. In most scenarios, this won't affect any
queries as the DECIMAL can auto-cast to DOUBLE; in the case were the
literal stands alone, the output schema will be a DECIMAL instead of a
DOUBLE. To specify a DOUBLE literal, use scientific notation (e.g.
1.234E-5).