chore: get push queries working with windowed primitive keys (#4369)
* chore: get push queries working with windowed primitive keys

The main issue here was that things like `DataSourceNode` and `StepSchemaResolver` were incorrectly copying `ROWKEY` into the value schema as-is for windowed keys. This was out of line with what the transformer for windowed keys was doing in `SourceBuilder`. The transformer was converting the windowed key into a `STRING` in the form: `<key> : Window:{start=<start>, end=<end>}`. Hence the SqlType of the `ROWKEY` in the value schema should be `STRING` for windowed sources.

NOTE: this means that the type of `ROWKEY` can not be different in the key and value schema, e.g. a windowed source with a `BIGINT` key will have `ROWKEY BIGINT KEY` as the key schema, but the copy of `ROWKEY` added to the source schema as it passes through the Streams topology will be `ROWKEY STRING` to reflect that fact its stored as a STRING, e.g. `"11 :  Window:{start=149532753000, end=-}"`.

This disconnect between the Sql types of `ROWKEY` in the key and value schema should be temporary. Once we support truly structured keys this will go.

This commit makes this change.

* `LogicalSchema.withMetaAndKeyColsInValue` now takes a `boolean` to indicate if the source is windowed or not.
* As the type of `ROWKEY` can not be different in the key and value schem `ExpressionTypeManager` needs to get the type from the value schema for push queries (or actually anything not a pull query).