fix: reserve `WINDOWSTART` and `WINDOWEND` as system column names
BREAKING CHANGE: `WINDOWSTART` and `WINDOWEND` are now reserved system column names. Any query that previously used those names will need to be changed: for example, alias the columns to a different name.

These column names are being reserved for use as system columns when dealing with streams and tables that have a windowed key.