Reorder value schema (#4412)
* refactor: re-order value schema used in Streams processing

Previously, when copying the meta and key columns into the value schema to enable processing via a Streams topology the columns where copied to the _start_ of the value schema. For example, given a schema of:

```sql
ROWKEY INT KEY, VAL0 STRING
```

The schema used during processing would of been:

```sql
ROWKEY INT KEY, ROWTIME BIGINT, ROWKEY INT, VAL0 STRING
```

Note: `ROWTIME` and `ROWKEY` columns are _before_ the existing value columns.

This meant the transformer in `SourceBuilder` had to add the value for `ROWTIME` and `ROWKEY`, (and now `WINDOWSTART` and `WINDOWEND` for windowed sources), to the _start_ of the list of values in the `GenericRow`.  Internally, `GenericRow` was using an `ArrayList`. Inserting at the start of an `ArrayList` requires all the existing elements to be shuffled down to make room for the new element, hence processing incurred an array copy operation when inserting each of these system columns.

Additionally, the buffer backing the `ArrayList` would first need to be resized to accomodate the new element, resulting in an allocation and another array-copy op.

Here's the transformer code from `SourceBuilder` and details of the operations happening:

```java
@Override
public GenericRow transform(final K key, final GenericRow row) {
  if (row == null) {
    return row;
  }

  // The call below:
  //   - allocates a new buffer of <current-size> + 1 to hold the new element
  //   - performs an array copy to move all existing elements down one slot
  //   - inserts ROWTIME at the start of the list.
  row.getColumns().add(0, processorContext.timestamp());

  // The call below:
  //   - allocates a new buffer of <current-size> + <num-key-columns> to hold the new elements
  //   - performs an array copy to move all the existing elements, except ROWTIME, down by <num-key-columns> slots
  //   - inserts the key columns after ROWTIME.
  row.getColumns().addAll(1, rowKeyGenerator.apply(key));
  return row;
}
```

As you can see this results in two new allocations and two array copies. Given this is in the main processing loop these aren't cheap and are easily avoided.

The new code does this:

```java
@Override
public GenericRow transform(final K key, final GenericRow row) {
    if (row == null) {
        return row;
    }

    final long timestamp = processorContext.timestamp();
    final Collection<?> keyColumns = rowKeyGenerator.apply(key);

    // Ensures capacity is large enough for all inserts
    row.ensureAdditionalCapacity(1 + keyColumns.size());
    // Appends ROWTIME at the end of the list
    row.getColumns().add(timestamp);
    // Appends key columns at the end of the list
    row.getColumns().addAll(keyColumns);
    return row;
}
```

This has zero-or-one allocations, (see below for why its generally zero), and not unnecessary array copies.

To avoid the allocation of a new buffer when the transformer calls `row.ensureAdditionalCapacity(1 + keyColumns.size());` there is a change in `GenericRowSerDe` to create the initial `GenericRow` with enough spare capacity to hold `ROWTIME`, `ROWKEY`, `WINDOWSTART` and `WINDOWEND`.

BREAKING CHANGE: the order of columns for internal topics has changed. The `DELIMITED` format can not handle this in a backwards compatible way. Hence this is a breaking change for any existing queries the use the `DELIMITED` format and have internal topics.

This change has been made now for two reasons:
 1. its a breaking change, making it much harder to do later.
 2. The recent  https://github.com/confluentinc/ksql/pull/4404 change introduced this same issue for pull queries. This current change corrects pull queries too.