fix: support conversion of STRING to BIGINT for window bounds
Fixes: https://github.com/confluentinc/ksql/issues/4482

Comparison and Between expressions in the WHERE clause already support magic conversion from a STRING containing a ISO formatted datetime into a BIGINT for the `ROWTIME` column.

This changes extends the support to cover the `WINDOWSTART` and `WINDOWEND` columns.

The change also fixes a bug where by a numeric Between expression on `ROWTIME` resulted in a class-cast exception, e.g. `WHERE ROWTIME < 123546794894`.