chore: include support metrics in the project (#4474)
ksqlDB is the only user of these classes, so include it in the project instead of consuming it as a library.

Source files were copied from Confluent Kafka's support-metrics-common module. They were updated to match ksqlDB checkstyle and to remove functionality that submits metrics to a Kafka topic (this was never used by ksqlDB).