chore: if provided, always use revision param in CI release job (#4644)
For ksql-db-release job parameters, currently any GIT_REVISION is ignored unless RELEASE_BUILD is true. For pre-release validation, we should be able to run a non-release job at a given sha. The principal thing we want to avoid is creating an unnecessary tag, which happens when RELEASE_BUILD is true.

This change allows the provided revision to also be used when RELEASE_BUILD is false.