refactor: do not allow duplicates in schemas by default
KsqlDB does not support data sources with duplicate column names in the key and value.  (Mainly due to the fact that it copies the key columns into the value while processing with Kafka Streams).

It therefore makes sense, to me at least, that our `LogicalSchema` rejects duplicates by default, but allows them if explicitly told to do so.  Doesn't change anything functionally, but does mean code is more explicit about when it does and does not allow duplicates.