test: avoid NPE if user puts input topic name in output
Fixes an NPE when a user accidentally uses an input topic name in the output records section of a test case, e.g.

```json
{
      "name": "ooops - cut & paste error",
      "statements": [
        "CREATE STREAM INPUT (a DECIMAL(4,2), b DOUBLE) WITH (kafka_topic='input', value_format='AVRO');",
        "CREATE STREAM OUTPUT AS SELECT (a + b) AS RESULT FROM INPUT;"
      ],
      "inputs": [
        {"topic": "input", "key": "0", "value": {"A": "10.01", "B":  5.1}}
      ],
      "outputs": [
        {"topic": "input", "key": "0", "value": {"RESULT": 15.11}},
        {"topic": "TEST2", "key": "0", "value": {"RESULT": 5.01}},
        {"topic": "TEST2", "key": "0", "value": {"RESULT": 10.01}}
      ]
    }
```

Note: the reference to `input` in the `outputs`? This was causing an NPE. It now causes an error along the lines of:

```
Topic `input`: expected 1 message, got 0.
```