chore: refactor AggregateAnalyzer
This is a follow on task from https://github.com/confluentinc/ksql/pull/4781. While working on #4781 I realized it the complex logic in `enforceAggregateRules` could be simplified by applying the rules along the way, rather than capturing all the information along the way and then trying to munge it all together at the end.  This simplifies the code and moves several fields from `MutableAggregateAnalysis` to `AggregateAnalyzer.AggAnalyzer`.