fix: do not filter out rows where PARTITION BY resolves to null
Fixes: https://github.com/confluentinc/ksql/issues/4747

This commit removes the filter that was excluding any rows where the `PARTITION BY` clause resolved to a `null` value, i.e. either because the result was `null` or because an error occurred evaluating the expression.

This change will only affect new queries started. Pre-existing queries will continue to run as before.