test: enhance QTT's historical tests to capture post conditions
The post conditions in a test are there to test certain conditions are true after the test case has run. They are defined inline within the QTT JSON tests.

Previously, the historical plans captured did not include the post conditions. Instead, the post conditions for the historical test cases were copied from the current test case. This means that any change in functionality that results in a change to the post-state of the system will cause any historic plans that test this state to fail.

For example, let's say we have a test case that partitions a stream by column `A`.  The historic plan may include a post condition that includes the schema of the result, which would have a key column named `ROWKEY`.  Now lets say we introduce a new Partition By step that instead creates a sink schema where the key column is called `A` and we update the test case to reflect this.  The historic plans will fail because the schema correctly has a key called `ROWKEY`, but the post condition wants it called `A`.

This commit changes the process so that the post conditions are included in the historical plans, so that the historical test cases test against the correct post conditions.