fix: Stop PARTITION BY and UDTFs that fail from terminating the query (#4822)
* fix: PARTITION BY and UDTFs that fail from terminating the query

`PARTITION BY` was recently enhanced to support any expression, rather than just column references.  If the expression expression results in an error, e.g. a UDF that throws an exception, this exception was not being handled - resulting in the query being terminated.

UDTFs were recently introduced. If a UDTF throws an exception, or an exception occurs extracting a UDTFs parameters, e.g. if a parameter is a UDF that throws, then the exception was not being handled - resulting in the query being terminated.

All these exceptions are now handled by logging the error to the processing logger and continuing.

Co-authored-by: Andy Coates <big-andy-coates@users.noreply.github.com>