refactor: correct optionals and remove empties from query plans (#4879)
* refactor: correct optionals and remove empties from query plans

This commit cleans up some properties that are marked both `required` and of type `Optional`, and others that are collections that are `required` but can be empty, in the types used to create ksqlDB's execution plans.

These changes remove a lot of noise from the plans, e.g. properties set to `null` or to empty collections `[]` and `{}`.

The saved historical plans have not be re-written. These changes will only be applied going forward and are backwards compatible.


Co-authored-by: Andy Coates <big-andy-coates@users.noreply.github.com>