chore: correct key column names in GROUP BY key struct (#4938)
quick refactor of `GroupByParamsFactory` while correcting the name of the key column. The two inner classes were incorrectly using `ROWKEY` as the key column name, not the name of the key in the schema, which can now be things other than `ROWKEY`. This wasn't causing any issues (yet) as the key column name is not persisted (yet).  But good to clean it up.

Co-authored-by: Andy Coates <big-andy-coates@users.noreply.github.com>