chore: decouple CLI from ROWKEY (#4929)
* chore: decouple CLI from ROWKEY

fixes: https://github.com/confluentinc/ksql/issues/4928

The server response when describing a source schema now includes `fieldType='SYSTEM'` or `fieldType='KEY'` to differentiate system columns, i.e. ROWTIME, and key columns from value column.

The default is no field type. This design decision was taken as the `FieldInfo` pojo is used not just to describe the columns in the schema, but also fields in structs and a fieldType of `VALUE` wouldn't make much sense for a struct field. Where as no field type kind of works.

Co-authored-by: Andy Coates <big-andy-coates@users.noreply.github.com>