refactor: avoid returning null (#4980)
Quick refactor to avoid returning null from `getAllLocalStorePartitionLags`.  In general the KSQL code does not return nulls. These means as someone working on the code you don't need to constantly check if the method you're calling can return null, and you don't need to litter the code with null checks. Things are also much less likely to fail with NullPointerExceptions.

In this case, the method can simply return an empty map, and the result is the same. Where a distinction is needed between a result and no result, then return `Optional<>`.

Co-authored-by: Andy Coates <big-andy-coates@users.noreply.github.com>