chore: remove ROWKEY from datagen (#4934)
This commit removes any use of `ROWKEY` from the DataGen app. This is being done as part of the work to allow any key column names, (https://github.com/confluentinc/ksql/issues/3536).

This is a compatible change as the key column name is not persisted in the data sent to Kafka. Hence the output doesn't change.

Co-authored-by: Andy Coates <big-andy-coates@users.noreply.github.com>