chore: testCase's should only store topics explicitly defined in the json
Fixes an issue where a historic test would inherit the wrong set of topics from its parent test case.

The `TestCase` was being built with a list of `Topic`s the test case used. This includes not just topics explicitly defined in the test case itself, but also those gleaned from parsing the statements and from input and output records.

When a historical plan was being used to build a new test case it was inheriting these topics, but these topics can be wrong!  Hence the `TestCase` (and RQTT equivalent) only stores the explicit topics and we delay adding in the three other sources until we need to initialize the list of topics.