feat: remove restriction on key column name
fixes: https://github.com/confluentinc/ksql/issues/3536

Previously, stream `KEY` columns and table `PRIMARY KEY` columns had to be named `ROWKEY`. This change removes this restriction. Key columns can now have _any_ name.

For example, a table of users with the `BIGINT` user id stored in the Kafka record's key may have previously been defined as:

```sql
CREATE TABLE USERS (ROWKEY BIGINT PRIMARY KEY, NAME STRING)
   WITH('KAFKA_TOPIC'='users', 'VALUE_FORMAT'='JSON');
```

With this change, this can now be defined with a more appropriate name for the table's primary key:

```sql
CREATE TABLE USERS (ID BIGINT PRIMARY KEY, NAME STRING)
   WITH('KAFKA_TOPIC'='users', 'VALUE_FORMAT'='JSON');
```

To get around the `ROWKEY` column name requirement, previous versions of ksqlDB allowed an alias to the key to be defined in the WITH clause, for example:

```sql
CREATE TABLE USERS (ROWKEY BIGINT PRIMARY KEY, NAME STRING, ID BIGINT)
   WITH('KEY'='ID', 'KAFKA_TOPIC'='users', 'VALUE_FORMAT'='JSON');
```

The downside of this approach is that it required a _copy_ of the key column in the value of the Kafka record, which was often not present. This is no longer required. The above statement can now also be simplified to:

```sql
CREATE TABLE USERS (ID BIGINT PRIMARY KEY, NAME STRING)
   WITH('KAFKA_TOPIC'='users', 'VALUE_FORMAT'='JSON');
```