Any key name (#5093)
* feat: remove restriction on key column name

fixes: https://github.com/confluentinc/ksql/issues/3536

Previously, stream `KEY` columns and table `PRIMARY KEY` columns had to be named `ROWKEY`. This change removes this restriction. Key columns can now have _any_ name.

For example, a table of users with a `BIGINT` user id stored in the Kafka record's key may have previously been defined as:

```sql
CREATE TABLE USERS (ROWKEY BIGINT PRIMARY KEY, NAME STRING)
   WITH('KAFKA_TOPIC'='users', 'VALUE_FORMAT'='JSON');
```

With this change, this can now be defined with a more appropriate name for the table's primary key:

```sql
CREATE TABLE USERS (ID BIGINT PRIMARY KEY, NAME STRING)
   WITH('KAFKA_TOPIC'='users', 'VALUE_FORMAT'='JSON');
```

To get around the `ROWKEY` column name requirement, previous versions of ksqlDB allowed an alias to the key to be defined in the WITH clause, for example:

```sql
CREATE TABLE USERS (ROWKEY BIGINT PRIMARY KEY, NAME STRING, ID BIGINT)
   WITH('KEY'='ID', 'KAFKA_TOPIC'='users', 'VALUE_FORMAT'='JSON');
```

The downside of this approach is that it required a _copy_ of the key column in the value of the Kafka record, which was often not present. This is no longer required. The above statement can now also be simplified to:

```sql
CREATE TABLE USERS (ID BIGINT PRIMARY KEY, NAME STRING)
   WITH('KAFKA_TOPIC'='users', 'VALUE_FORMAT'='JSON');
```

BREAKING CHANGE:

Statements containing PARTITION BY, GROUP BY or JOIN clauses will now produce different output schemas.

For PARTITION BY and GROUP BY statements, the name of the key column in the result is determined by the PARTITION BY or GROUP BY clause:

1. Where the partitioning or grouping is a single column reference, then the key column will have the same name as this column. For example:

```sql
CREATE STREAM OUTPUT AS 
   SELECT * FROM INPUT GROUP BY X;
-- OUTPUT will have a key column called X;
```

2. Where the partitioning or grouping is a single struct field, then the key column will have the same name as the field. For example:

```sql
CREATE STREAM OUTPUT AS 
   SELECT * FROM INPUT GROUP BY X->field1;
-- OUTPUT will have a key column called FIELD1;
```

3. Otherwise, the key column name will be system generated and be in the form `KSQL_COL_n`, where `n` is some positive integer.

In all case, except where grouping by more than one column, the new key column's name can be set by defining an alias in the projection. For example:

```sql
CREATE TABLE OUTPUT AS 
   SELECT USERID AS ID, COUNT(*) FROM USERS GROUP BY ID;
-- OUTPUT will have a key column named ID.
```

For groupings of multiple expressions there is currently no way to provide a name for the system generated key column. This is a shortcoming that will be fixed shortly when ksqlDB supports more than just a single key column.

For JOIN statements, the name of the key column in the result is determined by the join criteria.

1. For INNER and LEFT OUTER joins where the join criteria contains at least one column reference the key column will be named of the left most source whose join criteria is a column reference. For example:

```sql
CREATE TABLE OUTPUT AS 
    SELECT * FROM I1 JON I2 ON abs(I1.ID) = I2.ID JOIN I3 ON I2.ID = I3.ID;
-- OUTPUT will have a key column named I2_ID.
```

The key column can be given a new name, if required, by defining an alias in the projection. For example:

```sql
CREATE TABLE OUTPUT AS
   SELECT I2.ID AS ID, I1.V0, I2.V0, I3.V0 FROM I1 JON I2 ON abs(I1.ID) = I2.ID JOIN I3 ON I2.ID = I3.ID;
-- OUTPUT will have a key column named ID.
```

2. For FULL OUTER joins and other joins where the join criteria is not on column references, the key column in the output is not equivalent to any column from any source. The key column will have a system generated name in the form `KSQL_COL_n`, where `n` is a positive integer. For example:

```sql
CREATE TABLE OUTPUT AS
    SELECT * FROM I1 FULL OUTER JOIN I2 ON I1.ID = I2.ID;
-- OUTPUT will have a key column named KSQL_COL_0, or similar.
```

The key column can be given a new name, if required, by defining an alias in the projection. A new UDF has been introduced to help define the alias called `JOINKEY`. It takes the join criteria as its parameters. For example:

```sql
CREATE TABLE OUTPUT AS
    SELECT JOINKEY(I1.ID, I2.ID) AS ID, I1.V0, I2.V0 FROM  I1 FULL OUTER JOIN I2 ON I1.ID = I2.ID;
-- OUTPUT will have a key column named ID.
```

`JOINKEY` will be deprecated in a future release of ksqlDB once multiple key columns are supported.

Co-authored-by: Andy Coates <big-andy-coates@users.noreply.github.com>