fix: always represent double/decimal as decimal in qtt
This patch changes how we represent decimal/double in our qtt to always
represent these as decimals internally (e.g. in the Record/RecordNode).
Storing them as doubles can result in the scale not being preserved for
decimals. This mostly works pretty seamlessly, except when writing
kafka-serialized doubles to source topics. In this case, we have to
convert the value to a double before passing to the serializer.

With this change, we can also re-enable the trailing zeroes qtt