refactor: unify TransientQuery metadata (#5351)
* refactor: unify TransientQuery metadata

A change went in a while ago that created a transient query without creating `TransientQueryMetadata`. Instead is inline-subclassed the abstract `QueryMetadata` class, and replicated the logic for handling limits.

The direct subclassing of `QueryMetadata` is not ideal. All transient queries should be `TransientQueryMetadata`.

The replication of the limit handling logical is also not ideal.

This change fixes both of these issues.

It also fixes a but where the server would return a server error if the user used a limit of zero.

Co-authored-by: Andy Coates <big-andy-coates@users.noreply.github.com>